+++ CheMPS2/CASSCFnewtonraphson.cpp
@@ -84,7 +84,7 @@ double CheMPS2::CASSCF::solve( const int Nelectrons, const int TwoS, const int I
    // Determine the maximum NORB( irrep ) and the max_block_size for the ERI orbital rotation
    const int maxlinsize      = iHandler->getNORBmax();
    const long long fullsize  = ((long long) maxlinsize ) * ((long long) maxlinsize ) * ((long long) maxlinsize ) * ((long long) maxlinsize );
-   const string tmp_filename = CheMPS2::defaultTMPpath + "/" + CheMPS2::DMRGSCF_eri_storage_name;
+   const string tmp_filename = tmp_folder + "/" + CheMPS2::DMRGSCF_eri_storage_name;
    const int dmrgsize_power4 = nOrbDMRG * nOrbDMRG * nOrbDMRG * nOrbDMRG;
    // For ( ERI rotation, update unitary, block diagonalize, orbital localization )
    const int temp_work_size = (( fullsize > CheMPS2::DMRGSCF_max_mem_eri_tfo ) ? CheMPS2::DMRGSCF_max_mem_eri_tfo : fullsize );
@@ -221,7 +221,7 @@ double CheMPS2::CASSCF::solve( const int Nelectrons, const int TwoS, const int I
 
          assert( OptScheme != NULL );
          for ( int cnt = 0; cnt < dmrgsize_power4; cnt++ ){ DMRG2DM[ cnt ] = 0.0; } // Clear the 2-RDM ( to allow for state-averaged calculations )
-         DMRG * theDMRG = new DMRG( Prob, OptScheme );
+         DMRG * theDMRG = new DMRG( Prob, OptScheme, CheMPS2::DMRG_storeMpsOnDisk, tmp_folder );
          for ( int state = 0; state < rootNum; state++ ){
             if ( state > 0 ){ theDMRG->newExcitation( fabs( Energy ) ); }
             Energy = theDMRG->Solve();